coenocline returns a matrix with attribute on locations
This is a suggestion: do not change the UI, but let coenocline()
still return a simple matrix, but with attribute on sampling
locations, and use those in plotting.

This also means that you do not need to have functions like head,
tail, dim because standard R functions work. The simulations()
function is also unnecessary.

This also fixes a bug in lines method: you need locations or
lines are drawn in the order of the data.